requirements.txt
alembic==1.7.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.3
cryptography==36.0.2
Flask==2.0.3
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.1
importlib-resources==5.4.0
itsdangerous==2.1.0
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.1.0
pycparser==2.21
PyMySQL==1.0.2
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.19.2
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.31
urllib3==1.26.8
uuid==1.30
uWSGI==2.0.20
Werkzeug==2.0.2
zipp==3.7.0
